requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
beautifulsoup4
requests
openai
aiohttp==3.11.11
pillow==10.4.0
-e .